# Make runtests about 2x faster, by using parallelism continuously