@gnprice 2020-12-30
> The PR is split into two commits for easier review:

I'd also be happy to split these into two PRs, if that'd be easier for your workflow. (If I were using Phabricator here, I'd probably make this a stack of two diffs.)