@panagosg7 2021-01-19
I was wondering how the unit tests suddenly got so much faster. Thanks for the awesome perf win @gnprice !